Move fake_config to mocks directory
