+++ test/mocks/fake_config.go
@@ -12,7 +12,7 @@
  limitations under the License.
 */
 
-package config
+package mocks
 
 import (
 	"fmt"
